Introducing nox
