+++ mail_groene_maaiers/groene_maaiers.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 """Groene maaiers script for sending emails to enlisted users on a Gsheet."""
-import os
 import re
 import smtplib
 import ssl
@@ -12,7 +11,6 @@ from email.message import EmailMessage
 
 from apiclient import discovery
 from decouple import config
-from google.oauth2 import service_account
 from google.oauth2.service_account import Credentials
 from rich import print as pprint
 
@@ -348,39 +346,3 @@ class ScheduleSheet(GSheet):
 
         # Should send admin email
         return [], "No names found."
-
-
-def main() -> None:
-    """Call main function."""
-    base_path = os.path.dirname(os.path.abspath(__file__))
-    credentials_file = os.path.join(base_path, "credentials.json")
-    scopes = [
-        "https://www.googleapis.com/auth/contacts.readonly",
-        "https://www.googleapis.com/auth/spreadsheets.readonly",
-    ]
-    credentials = service_account.Credentials.from_service_account_file(
-        credentials_file, scopes=scopes
-    )
-    notify = EmailNotification()
-    schedule_sheet = ScheduleSheet(credentials=credentials, notification=notify)
-    schedule_sheet.get_sheet()
-    names, err = schedule_sheet.names_next_date()
-    if schedule_sheet.date_not_found:
-        return
-    if err:
-        notify.admin_message(err)
-        notify.send_message()
-        return
-
-    # continue matching it with the contact information
-    contacts = Contacts(credentials=credentials, notification=notify)
-    contacts.get_sheet()
-    contacts.get_contact_name_email()
-    mailing_list = contacts.generate_mailing_list(names)
-
-    notify.standard_message(names=names, emails=mailing_list)
-    notify.send_message()
-
-
-if __name__ == "__main__":
-    main()
