+++ run
@@ -44,12 +44,29 @@ function test {
 	_poetry "ruff ."
 	_poetry "mypy . --junit-xml mypy_report.xml"
 	_poetry "black --diff --color ."
-	_poetry "pytest"
+	_poetry "pytest --cov-report=html"
 }
 
 function fix {
 	_poetry "black ."
 	_poetry "ruff . --fix"
+}
+
+function nox {
+  # -r, re-used existing virtualenv
+	_poetry "nox -r"
+}
+
+function nox:tests {
+	_poetry "nox -r -s tests"
+}
+
+function nox:lint {
+	_poetry "nox -r -s lint"
+}
+
+function nox:typing {
+	_poetry "nox -r -s typing"
 }
 
 function build {
